AI-Service/requirements.txt
fastapi
uvicorn
pymongo
pandas
pydantic
python-dotenv
scikit-learn
httpx
